Moors now spawn in Grenada
Slightly more historical capital than Cadiz.
Move Cadiz crab closer to Portugal.
Increases the urban density of Andalusia, which is a deliberate move to lower the efficiency of each city in that area.
Add fish and copper near Plymouth and encourage England to settler there.
Drain Norwich marsh in 1630. It took longer to completely drain it, but this makes sense for gameplay purposes.
Create a gap in Vietnamese respawn period between 1700 and 1945
Increase Vietnam resurrection probability per turn to 8.