Chiang Kai Shek now prefers Protestantism
Chiang Kai Shek is now leaderhead if Wu respawns after 1830
Add event to spread Protestantism in South China and Korea in the 19th and 20th centuries.
Add event to spread Catholicism in Japan in 1550-1700 but only if Japan isn't Isolationist.
Update religion maps to add Catholicism as minority in Japan and Protestantism as minority in South China.