Ghurid/Ghaznavid capital now changes to Delhi on conquest
Give TImurids some extra starting techs.
Add Quba spawn in 1350, so that if the Caucasus is empty, there is at least one city for the Timurids to conquer for their UHV.
Add Mosul to Spanish name map.
Remove "by 1660" requirement for Timurid wonders; the race should be against other civs to build the wonders.
Fix missing text key when Timurids in Mughal phase.