Move wine 1N of Aleppo
Qin should not start with a civic it doesn't have the tech for.
Increase priority of Firearms tech for Ottomans.
Slight buff to Ottoman, French modifiers.
Redefine arquebusiers default AI as city defence to the AI is encouraged to upgrade crossbowmen to Arquebusiers.
Tweak tech priorities for a few civs.
Add Potato south of Berlin in 1700 for the Prussian spawn.