Convert all dynamic names, adjectives, titles, leaders, languages and leader names from if-else to a dictionary
This should be both easier to read and modify, as well as (in theory) more performant.
Set iNextTurnAIWar from -600 to -2000 in 3000 BC starts.
Delete obsolete Barbs.py
Optimize a few BeginGameTurn functions to early exit.
Do not run script that re-fragments independents. They are already decently fragmented without that script.
Reduce the number of events that cause all of a civ's cities to be renamed. This won't be much of a visible change, just an optimization.
Adjust a war preference modifier that makes Spain more warlike against Moors OR Morocco and not just Moors.
Increase interval of minor cities getting an extra unit from 10 to 12
increaseAggressionLevels every 5 turns for all civs instead of every turn. Quintuple the possible aggression gain.