Change Dravidia respawn period to (840, 1650)
Byzantium no longer starts with Lateen Sails
Add barbarian invaders for Tamils in the 300-800 period. Only an issue if there is unclaimed land or barbarian land between India and Dravidia.
Move Berlin site (Brandenburg in HRE era) 1 west.
Adjust Austrian and German settler maps to prioritize this site and not the old Berlin, or Magdeburg
Add Modern German language for Brandenburg -> Berlin & Regensburg -> Munich
Extend HRE flip zone a few tiles north, to get some of the tiles of barbarian Stettin
Remove the Ghurid resurrection period of 1220-1400
Add a respawn period for HRE between 1050 and 1350
Restrict Byzantine invasion of Spain to the South-East (not Barcelona)