Split Cologne & Aachen cities
Aachen is now 2 west of Cologne and is its own city site, the capital of the Franks. This improves the division of the Carolingian Empire and is more favourable to the HRE, to protect it from French expansionism.
Require Shia founded for Assyria respawn to be called Hamdanid/Antiochene
This fixes the case where Assyria respawns in the pre-Islamic era.
Celts now uses Latin & Greek as fallback languages.