Add barbarian Danzig prior to spawn of Poland
This increases likelihood that Prussia will get 2 cities.
Chiang Kai Shek is less likely to vassalize.
Remove sheep from Azores (it causes Portugal to not want to settle the island for some reason).
Fix bug with conquest events, such that if flagged as "only conquer cities from preferred target", it does not target other cities. Currently, it would fire as long as the target civ existed, even if it didn't own cities in the target area.